test: TT-356 Adding azure endpoint api test
