+++ V2/azure_time_tracker/time_tracker/tests/conftest.py
@@ -1,0 +1,1 @@
+from time_tracker.tests.api.api_fixtures import activities_json
